+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Builtin function `read_entire_file(filename: String)`
+- String indexing
 
 ### Changed
 ### Fixed